ingest/requirements.txt
qdrant-client==1.9.1
requests==2.32.3
pypdf==4.3.1
python-docx==1.1.2
Pillow==10.4.0
markdown==3.7
tqdm==4.66.5

